backend/requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18          # file uploads

# ── Settings ─────────────────────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.6.0

# ── AWS / Bedrock ─────────────────────────────────────────────────────────────
boto3>=1.35.0
langchain-aws>=0.2.0

# ── LangChain core ────────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# ── Vector store ──────────────────────────────────────────────────────────────
langchain-chroma>=0.1.4
chromadb>=0.5.0

# ── PDF extraction ────────────────────────────────────────────────────────────
unstructured[pdf]>=0.16.0
pypdf>=5.0.0
pillow>=11.0.0
opencv-python-headless>=4.10.0
pdfplumber>=0.11.0
tiktoken>=0.8.0

# ── RAGAS evaluation ──────────────────────────────────────────────────────────
ragas>=0.2.0
datasets>=3.1.0

# ── Dev / test ────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0                     # async test client
